pyproject.toml
[project]
name = "projectai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2",
    "pydantic-settings>=2",
    "SQLAlchemy>=2",
    "alembic>=1.13",
    "requests>=2",
    "psycopg>=3.2.10",
    "black>=25.9.0",
    "psycopg-binary>=3.2.10",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "scikit-learn>=1.7.2",
    "google-generativeai>=0.8.5",
]

[project.scripts]
wb-ingest = "app.cli.ingest:main"
wb-worker = "app.workers.generate_worker:main"


[tool.alembic]
script_location = "alembic"
